% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_gen.R
\name{palette_gen}
\alias{palette_gen}
\title{Generate SRHD colrs for use in a chart.}
\usage{
palette_gen(palette = "main", direction = 1)
}
\description{
"palette_gen" identifies the number of colors to be used for 
discrete variables
}
\details{
Thisis a helper files to build "scale_*_srhd" functions for 
use in ggplot2. 


Method and code borrowed from here.
https://meghan.rbind.io/blog/2022-10-11-creating-custom-color-palettes-with-ggplot2/

Package creation instructions were from here.
https://hilaryparker.com/2014/04/29/writing-an-r-package-from-scratch/
}
