% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_srhd.R
\name{scale_color_srhd}
\alias{scale_color_srhd}
\title{Generate SRHD colors palette function for use in ggplot2.}
\usage{
scale_color_srhd(palette = "main", direction = 1, ...)
}
\description{
"scale_color_srhd" builds a function for use in ggplot2
that uses the SRHD color palette
}
\details{
This builds "scale_color_srhd" functions for 
use in ggplot2 using the SRHD color palette.
Color is the border or line.
}
