% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_srhd_c.R
\name{scale_color_srhd_c}
\alias{scale_color_srhd_c}
\title{Generate SRHD colors palette function for use in ggplot2.}
\usage{
scale_color_srhd_c(palette = "main", direction = 1, ...)
}
\description{
"scale_color_srhd_c" builds a function for use in ggplot2
that uses the SRHD color palette
This one is for continuous scale or gradient color
}
\details{
This builds "scale_color_srhd_c" functions for 
use in ggplot2 using the SRHD color palette.
It creates a two color gradient.
Color is the border or line.
}
