% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srhd_color.R
\name{srhd_color}
\alias{srhd_color}
\title{SRHD colors}
\usage{
srhd_color(...)
}
\description{
"srhd_color" names the hex colors to be used.
}
\details{
This color palette use Spokane Regional Health District colors from 2021.
These colors can be used to color charts.
This is also a helper file to build "srhd_palette", which in turn buildsthe 
"palette_gen" and "palette_gen_c" functions.
And those in turn are helper files to build "scale_*_srhd" functions for 
use in ggplot2. 

Method and code borrowed from here.
https://meghan.rbind.io/blog/2022-10-11-creating-custom-color-palettes-with-ggplot2/

Package creation instructions were from here.
https://hilaryparker.com/2014/04/29/writing-an-r-package-from-scratch/
#Example:
#use palette in a chart
str(iris)

#no color defined
iris %>% 
 count(Species) %>% 
 ggplot(aes(x = Species, y = n))+
 geom_bar(stat = "identity")
#the new srhd_color function is used to define a color
iris %>% 
 count(Species) %>% 
 ggplot(aes(x = Species, y = n))+
 geom_bar(stat = "identity", fill = srhd_color("lt green"))
 
#if you want to specify the colors use scale_*_manual
iris %>% 
 count(Species) %>% 
 ggplot(aes(x = Species, y = n, fill = Species)) +
 geom_bar(stat = "identity") +
 scale_fill_manual(values= unname(c(srhd_color(
 "teal", "orange", "dk blue"))))
}
