% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srhd_palette.R
\name{srhd_palette}
\alias{srhd_palette}
\title{SRHD color palette}
\usage{
srhd_palette(palette = "main", ...)
}
\description{
"srhd_palette" builds a list with one or more color palettes
}
\details{
These colors can be used to color charts.
This is a helper file to build "palette_gen"and "palette_gen_c" functions.
And those in turn are helper files to build "scale_*_srhd" functions for 
use in ggplot2. 

call the color palette using the palette function
srhd_palette("main")

use the show_col function from the scales package to show the actual color 
and not just hex codes
scales::show_col(srhd_palette("main"), cex_label=2)

Method and code borrowed from here.
https://meghan.rbind.io/blog/2022-10-11-creating-custom-color-palettes-with-ggplot2/

Package creation instructions were from here.
https://hilaryparker.com/2014/04/29/writing-an-r-package-from-scratch/
}
